added auth middleware to workout routes

+++ server/src/api/controllers/workouts/index.ts
@@ -1,9 +1,10 @@
 import Workout from '../../../models/workouts'
+import User from '../../../models/user'
 import {Request, Response} from 'express'
 
 const getWorkouts = async (req:Request, res:Response) => {
     try{
-        const workouts = await Workout.find()
+        const workouts = await Workout.find().populate("created_by", "name picture")
         res.status(200).json({status:200, log:"Workouts fetched successfully", workouts })
     }catch(e:any){
         res.status(500).json({ status: '500', log: e.message })
@@ -17,7 +18,7 @@ const getWorkoutById = async (req:Request, res:Response) => {
             res.status(400).json({ status: '400', log: "workout id required" })
             return
         }
-        const workout = await Workout.findOne({_id })
+        const workout = await Workout.findOne({_id }).populate("created_by", "name picture")
         res.status(200).json({status:200, log:"Workout fetched successfully", workout })
     }catch(e:any){
         res.status(500).json({ status: '500', log: e.message })
@@ -27,12 +28,20 @@ const getWorkoutById = async (req:Request, res:Response) => {
 const createWorkout = async (req:Request, res:Response) => {
     try{
         const {name, body_parts_targeted, notes, exercises} = req.body
+        // @ts-ignore
+        const {email} = req.user
+        const created_by = await User.findOne({email})
+        if(!created_by){
+            res.status(403).json({ status: '403', log: "Unauthorized, create an account first" })
+            return
+        } 
         // get created_by from auth middleware 
         const workout = new Workout({
             name, 
             body_parts_targeted,
             notes,
             exercises,
+            created_by : created_by._id
         })
         const doc = await workout.save()
         res.status(200).json({status:200, log:"Workout successfully created", workout: doc })
@@ -52,7 +61,17 @@ const deleteWorkoutById = async (req:Request, res:Response) => {
             res.status(400).json({ status: '400', log: "workout id required" })
             return
         }
-        const workout = await Workout.deleteOne({_id })
+        const workout = await Workout.findOne({_id}).populate("created_by", "email")
+        if(!workout){
+            res.status(400).json({ status: '400', log: "Workout not found" })
+            return
+        }
+        // @ts-ignore
+        if(workout.created_by.email !== req.user.email) {
+            res.status(403).json({ status: '403', log: "Unauthorized to delete this" })
+            return
+        }
+        await Workout.deleteOne({_id })
         res.status(200).json({status:200, log:"Workout deleted successfully" })
     }catch(e:any){
         res.status(500).json({ status: '500', log: e.message })
@@ -67,9 +86,19 @@ const updateWorkoutById = async (req:Request, res:Response) => {
         res.status(400).json({ status: '400', log: "workout id required" })
         return
     }
+    const workout = await Workout.findOne({_id}).populate("created_by", "email")
+    if(!workout){
+        res.status(400).json({ status: '400', log: "Workout not found" })
+        return
+    }
+    // @ts-ignore
+    if(workout.created_by.email !== req.user.email) {
+        res.status(403).json({ status: '403', log: "Unauthorized to update this" })
+        return
+    }
     await Workout.updateOne({_id}, args)
-    const workout = await Workout.findOne({_id})
-    res.status(200).json({status:200, log:"Workout updated successfully", workout})
+    const updatedWorkout = await Workout.findOne({_id})
+    res.status(200).json({status:200, log:"Workout updated successfully", workout:updatedWorkout})
     }catch(e:any){
         res.status(500).json({ status: '500', log: e.message })
     }
